fix: renamed folder and import names
