+++ frontend-next/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.24.1",
     "cookies-next": "^4.1.1",
     "moment": "^2.30.1",
     "next": "14.1.0",
