+++ frontend-next/src/repositories/auth/types.ts
